{{cookiecutter.repo_name}}/pyproject.toml
[tool.poetry]
name = "fastapi-genesis"
version = "0.1.0"
description = "FastAPI code Template"
authors = ["Luka Vuksanovic <luka.vuksanovic35@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
black = "^23.7.0"
alembic = "^1.11.1"
loguru = "^0.7.0"
flake8 = "^6.0.0"
docker-compose = "^1.29.2"
fastapi = {extras = ["all"], version = "^0.100.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
cookiecutter = "^2.2.3"
bump-pydantic = "^0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
